+++ shared/util/signature.py
@@ -2,8 +2,9 @@ import base64
 import hashlib
 import hmac
 import logging
+import textwrap
 import time
-from typing import Any, Literal
+from typing import Any, Final, Literal
 
 from aiohttp.typedefs import Query
 from fastapi import HTTPException, status
@@ -31,15 +32,16 @@ def get_signature_headers(
     logging.debug(
         f"Getting signature headers for: \n{method} \n{path} \n{body} \n{params}"
     )
-    timestamp = int(time.time())
-    headers = {X_TIMESTAMP: str(timestamp)}
-    if not secret_key:
-        return headers
 
-    signature = _get_req_signature(
-        secret_key, timestamp, method, path, body, params
-    )
-    headers[X_SIGNATURE] = signature
+    timestamp: Final[int] = int(time.time())
+    headers: Final[dict[str, str]] = {X_TIMESTAMP: str(timestamp)}
+
+    if secret_key:
+        signature: Final[str] = _get_req_signature(
+            secret_key, timestamp, method, path, body, params
+        )
+        headers[X_SIGNATURE] = signature
+
     logging.debug(f"Signature headers: {headers}")
     return headers
 
@@ -63,44 +65,66 @@ def verify_signature_headers(
     :param body: body of the req
     :param params: query parameters of the req
     """
-    current_timestamp = int(time.time())
-    timestamp = int(headers.get(X_TIMESTAMP, "0"))
-    signature = headers.get(X_SIGNATURE, "")
     logging.debug(
         f"Verifying signature headers for:\n{method}\n{path}\n{body}\n{params}\n{headers}"
     )
-    if abs(current_timestamp - timestamp) > signature_ttl:
-        message = f"""\
-Signature expired for:
-method={method}
-path={path}
-body={body}
-params={params}
-age={current_timestamp - timestamp}s
-current_timestamp={current_timestamp}s
-request_timestamp={timestamp}s
-"""
+
+    if not secret_key:
+        message = "AGENT_SECRET is not set, cannot verify request signature"
         logging.warning(message)
         raise HTTPException(
             status.HTTP_401_UNAUTHORIZED,
             message,
         )
-    if not secret_key:
-        return True
-    expected = _get_req_signature(
-        secret_key, timestamp, method, path, body, params
-    )
-    if not hmac.compare_digest(expected, signature):
-        message = f"""\
-Invalid signature for:
-method={method}
-path={path}
-body={body}
-params={params}
-signature={signature}
-"""
+
+    if not headers.get(X_TIMESTAMP):
+        message = f"{X_TIMESTAMP} header is missing, cannot verify request signature"
         logging.warning(message)
-        raise HTTPException(401, message)
+        raise HTTPException(
+            status.HTTP_401_UNAUTHORIZED,
+            message,
+        )
+
+    if not headers.get(X_SIGNATURE):
+        message = f"{X_SIGNATURE} header is missing, cannot verify request signature"
+        logging.warning(message)
+        raise HTTPException(
+            status.HTTP_401_UNAUTHORIZED,
+            message,
+        )
+
+    current_timestamp: Final[int] = int(time.time())
+    timestamp: Final[int] = int(headers.get(X_TIMESTAMP, "0"))
+    signature: Final[str] = headers.get(X_SIGNATURE, "")
+
+    if abs(current_timestamp - timestamp) > signature_ttl:
+        message = textwrap.dedent(f"""\
+            Signature expired for:
+            method={method}
+            path={path}
+            body={body}
+            params={params}
+            age={current_timestamp - timestamp}s
+            current_timestamp={current_timestamp}s
+            request_timestamp={timestamp}s""")
+        logging.warning(message)
+        raise HTTPException(
+            status.HTTP_401_UNAUTHORIZED,
+            message,
+        )
+
+    expected = _get_req_signature(secret_key, timestamp, method, path, body, params)
+    if not hmac.compare_digest(expected, signature):
+        message = textwrap.dedent(f"""\
+            Invalid signature for:
+            method={method}
+            path={path}
+            body={body}
+            params={params}
+            signature={signature}""")
+        logging.warning(message)
+        raise HTTPException(status.HTTP_401_UNAUTHORIZED, message)
+
     return True
 
 
@@ -126,9 +150,7 @@ def _get_req_signature(
 
 def _get_sig_encoded(secret_key: str, sig_bytes: bytes) -> str:
     return base64.b64encode(
-        hmac.new(
-            secret_key.encode(), sig_bytes, hashlib.sha256
-        ).digest()
+        hmac.new(secret_key.encode(), sig_bytes, hashlib.sha256).digest()
     ).decode()